Added a hardwired font to the code after chatGPT informed me that the GLX function "glXUseXFont()" was not guaranteed to be found in modern drivers, and Slackware64 15.0 was consistent with this. The font was taken form the Linux 8x8 console Cp850 font. The raw data in ASCII format in C were taken from executing "monobit" (using Python3) on the file:
/usr/share/kbd/consolefonts/cp850-8x8.psfu.gz

Flipping horizontally and capturing was done: "flip to --format=c"

The global (GLubyte) array "font_Cp850_8X8_8x8_flip" contains the data in
hexadecimal. They are parsed in 8 rows, of 1 byte each.

The function "xwindow_setupLinuxFont()" was added to "inxlib.c" to create
the OpenGL display lists, and some constants were adjusted by IN. Note that
including the first letter causes the Xorg driver to malfunction on my
system, while it works at others; skipping it appears to work as exected.

The driver in "test.c" was modified to use the newly added font.
IN 2023/05/16 "master"