Begun the transition to version v0.6 which is a major upgrade to the OpenGL handling. Several slight changes were made in anticipation for how the new version will function. (I held back on this commit for a month, so as to test the changes for bugs while testing with "programmable pipeline" at the same time.)
Changed to having 2 GL contexts, accessible via two variables, and there is
a name-change that had to be propagated to several functions for backward
compatibility.

Added the function "xwindow_setup_dualglx()" which is supposed to bring up
a window with two associated GL contexts, which are to share resources. This
has not yet been fully implemented, but the function properly brings up the
first context, which can be used to further develop before finalizing this
functionality. The option for the "old style" initialization is retained.
The second GL context remains unallocated and unuseable.

Added the callbacks "FrameEntry" and "FrameExit", which are meant to perform
operations prior to and after drawing a frame, and trapping events.
IN 2023/12/02 "master"