The code was augmented substantially in order to use both the fixed and the programmable pipelines for rendering and rasterization. The build is an example of using both; no GLUT is used, and the user is in complete control of everything, from opening a window, to trapping keyboard and mouse events, to rendering with both pipelines.
The "README.md" file was augmented with the "TL;DR" info, and it also has a
"crash course" on the programmable pipeline and how to use both pipelines.

The "dummy_draw.c" code was heavily modified to do the high level work of
drawing using the fixed and the programmable pipelines. The old one is used
to render text on the screen and a wireframe cube. The modern one is used to
draw a simple object made of 3 coloured triangles and a coloured sphere made
of many triangles. Inspection of the code (and the comments) demonstrates
how the two pipelines are used, one at a time. (Note that to maintain the
same projection and modelview, I extract the corresponding matrices from the
fixed pipeline and push them to the programmable one as "uniform variables".)
This code includes the "my_demo.c" file, which does all the drawing and the
setting up of the programmable pipeline -- including generating the objects.

The file "my_demo.c" has some structs, which are used to hold data in global
scope in order to simplify the process. It contains functions that create the
vertex and attribute buffer objects (VBO/VAO) for two types of objects. It
also sets up the programmable pipeline's shader program, and to do so it uses
code found in "inogl.*". This file contains the vertex and fragment shader
codes. (Studying this file will help understand the modern pipeline.) There
is a lot of error-trapping in those functions to aid in isolating errors when
making modifications. A function that is used to prepare the fixed pipeline
at each frame is also included. This file includes "axissphere.c", which is
a utility code to build a sphere object.

The file "inogl.h" shows all the structs that are used to handle a shader
program and VBO/VAOs associated with drawing operation of specific objects.
It also has samples and information on the vertex/fragment shaders. The file
"inogl.c" has functions that allow for setting up the OpenGL programmable
pipeline, creating shader programs (based on the specific program structs),
querying uniform variables, and setting up VAO/VBOs.

The driver "test.c" was modified to call the "init()" function that is found
in the file "my_demo.c". This is because initialization of the OpenGL has to
be done _after_ the GL context was created, and there is no callback that
does one-time operations (that is what drivers are for). This init() call is
what creates the objects that are to be rendered.
IN 2023/12/09 "master"