This is a simple change, but a substantial one. This changes the version of this library to v0.7 to reflect the change.
I have changed the file that was the "test.c" driver to be in the new file
"inxlib_gui.c", and all the "inxlib*" files are now compiled into the shared
object library file "libINXlib.so". (In this library I have added all user
code that would make up the application.) A driver called "test.c" was added
to perform the singlae function of (a) opening the shared object library in
the present directory, (b) locate the entry point function, and (c) execute
it using a local variable for the struct that handles "inxlib.*" data. What
this (monumental and yet simple) change allows is for a program that uses
this library as a GUI to _not_ be compiled against it. Although it does load
it explicitly in this example, imagine the possibilities... (I will leave it
at that.)
IN 2025/03/17 "master"